# gateco/packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SimulationRequest": {
        "properties": {
          "connector_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connector Id"
          },
          "principal_id": {
            "title": "Principal Id",
            "type": "string"
          },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          },
          "resource_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Ids"
          }
        },
        "required": [
          "principal_id"
        ],
        "title": "SimulationRequest",
        "type": "object"
      },
      "SimulationResponse": {
        "properties": {
          "allowed": {
            "title": "Allowed",
            "type": "integer"
          },
          "denial_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Denial Reasons",
            "type": "array"
          },
          "denied": {
            "title": "Denied",
            "type": "integer"
          },
          "matched_resources": {
            "title": "Matched Resources",
            "type": "integer"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          },
          "policy_trace": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Policy Trace",
            "type": "array"
          }
        },
        "required": [
          "outcome",
          "matched_resources",
          "allowed",
          "denied",
          "policy_trace",
          "denial_reasons"
        ],
        "title": "SimulationResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Backend API for gateco",
    "title": "gateco API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/simulator/run": {
      "post": {
        "description": "Dry-run policy evaluation for a principal against resources.",
        "operationId": "run_simulation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Run Simulation",
        "tags": [
          "simulator"
        ]
      }
    },
...
  }
}